Move SliceInput into conn::receive
