+++ rustls/src/msgs/mod.rs
@@ -57,7 +57,6 @@ pub(crate) use codec::{
 use codec::{LengthPrefixedBuffer, U24};
 
 mod deframer;
-pub use deframer::SliceInput;
 pub(crate) use deframer::{Deframed, Deframer, Delocator, HandshakeAlignedProof, Locator};
 
 mod enums;
